authentication

+++ my_project/crud/user.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException, status
 from database.models import User
 from schemas import UserBase
 from sqlalchemy.orm.session import Session
@@ -21,11 +22,16 @@ def get_all_users(db: Session):
 
 # Get User
 def get_user(db: Session, id: int):
-   return db.query(User).filter(User.id == id).first()
+   user = db.query(User).filter(User.id == id).first()
+   if not user:
+      raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+   return user
 
 # Update User
 def update_user(db: Session, request: UserBase, id: int):
    user = db.query(User).filter(User.id == id)
+   if not user.first():
+      raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
    user.update({
       User.username: request.username,
       User.email: request.email,
@@ -39,9 +45,8 @@ def update_user(db: Session, request: UserBase, id: int):
 # Delete User
 def delete_user(db: Session, id: int):
    user = db.query(User).filter(User.id == id).first()
+   if not user:
+      raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
    db.delete(user)
    db.commit()
    return {"message": "User Deleted Successfully!"}
-
-
-#---------------------------------------------------------------
